requirements.txt
python-telegram-bot
requests
mysql-connector-python
beautifulsoup4
lxml